remove fifo

+++ src/adc.h
@@ -9,4 +9,4 @@ void ADC_init(void);
 void ADC_down(void);
 inline void ADC_channel(uint8_t channel);
 inline uint16_t ADC_get(void);
-float getTemperature();
+unsigned int getTemperature();
